feat/#4: 학생 회원 가입 기능 추가

+++ src/main/java/inha/git/mapping/domain/id/PatentRecommedId.java
@@ -18,7 +18,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class PatentRecommedId implements Serializable {
 
-    private Long projectId;
-    private Long userId;
+    private Integer projectId;
+    private Integer userId;
 
 }
